requirements.txt
fritzconnection
python-eq3bt
paho-mqtt
requests
astral
parsel
redis
